Added README, updated image animators
Added a faster image animator script, and shifted the old one into
mpl_image_animator.py. Also added a README.
